requirements.txt
# Django и DRF
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Аутентификация и авторизация
django-allauth==0.57.0
dj-rest-auth==4.0.1

# WebSocket и Channels
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# База данных
psycopg2-binary==2.9.7
redis==5.0.1

# CORS и безопасность
django-cors-headers==4.3.1
django-filter==23.3

# Документация API
drf-spectacular==0.26.5

# Утилиты
python-decouple==3.8
Pillow==10.0.1

# Тестирование
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1

# Форматирование кода
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Разработка
django-debug-toolbar==4.2.0
